refactor: move component files to their folders
