+++ tests/components/CharacterDetails.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import '@testing-library/jest-dom';
-import CharacterDetails from '../../src/components/CharacterDetails';
+import CharacterDetails from '../../src/components/CharacterDetails/CharacterDetails';
 import React from 'react';
 
 declare const global: {
